Move istio installed check to client
