+++ src/controllers/dynakube/istio/config.go
@@ -1,6 +1,8 @@
 package istio
 
 import (
+	"fmt"
+
 	"github.com/Dynatrace/dynatrace-operator/src/logger"
 )
 
@@ -11,4 +13,11 @@ var (
 const (
 	OperatorComponent = "operator"
 	OneAgentComponent = "oneagent"
+
+	IstioGVRName    = "networking.istio.io"
+	IstioGVRVersion = "v1alpha3"
+)
+
+var (
+	IstioGVR = fmt.Sprintf("%s/%s", IstioGVRName, IstioGVRVersion)
 )
